utils.py: Remove check for symlinks that overwrite the host
This cannot happen in practice as we make sure that every symlink inside
an artifact is relative when we create the artifact.

The new integration-tests/symlink-test should ensure we don't regress there.

This saves one call to os.path.realpath() for every file we stage, which
makes a large performance difference.