doc: Promote documention on how to use external plugins up one level
Currently the front page at https://buildstream.gitlab.io/buildstream/
doesn't make it easy to discover how to use external plugins.

There is an 'External plugins' heading which simply links to the
bst-external documentation (which also doesn't document how to use
external plugins) and that is all.

By promoting the "Plugin Origins and Versions" item up a level, we
cause it to appear in the table of contents on the front page making
it easier to find.