Add test to catch conversion edge case
In cases where conversion is required for the workspaces.yml config, it
is the case that this is not being handled correctly in that the newly
loaded format would not contain provenance information, causing problems
for code that expects this provenance information, such as operations
like closing a workspace.

This commit introduces a test to pick up on this failure.