Change workspaces.yml conversion technique
Previously we had been generating a dict for the new version 1
workspaces.yml after a conversion, which did not have any provenance
information, causing failures down the line where provenance was
expected.

This change brings the conversion stage, which includes the persisting
of the converted workspaces.yml, into the loading step. This is to
ensure that provenance information is always available for the current
config on disk, rather than some transient copy.