BREAK: manual.yaml: don't set any default env vars
Remove the 'MAKEFLAGS' and 'V' defaults.

Now that there is a specialised 'make' element, these make-specific
defaults don't make sense here. This element is meant to be a blank
slate for folks to build on.

Note that this is a breaking change, that will require some users to
make changes to their .bst files if they are expecting these environment
variable to be set.

_versions.py: BST_FORMAT_VERSION bumped to 18.