element.py: Introduce __update_ready_for_runtime()
An element becomes ready for runtime once it has a
cache key and all of its transitive runtime dependencies
have cache keys.

This patch introduces the method __update_ready_for_runtime()
which is called once the cache key of an Element is calculated.
When called, we immediately notify reverse dependencies and,
if appropriate, try to calculate their cache keys.