Stream: Fix the existence of duplicate queues
It was possible for multiple Queues of the same type to exist.
Currently, there is no desired reason for this to happen.

These changes add an explicit function call to the Scheduler that
destroys the queues, to be used before constructing the next list of
queues to pass into the Scheduler.

It also calls this in all the places before we construct the queues.

Further, it alters Stream.fetch_subprojects because there is currently
no reason why we'd want to preserve the Stream's list of queues before
running.