Remove the suggestion that BuildGrid can be used as an artifact server
BuildGrid's CAS can be used as artifact storage, but BuildGrid no longer
supports being used as an Artifact Cache, since it doesn't implement the
new proto-based approach to artifacts.

This commit makes this current situation clear in the docs, rather than
incorrectly suggesting the use of BuildGrid for the `index` and `both`
cache types.